checking and working
